Ahora ya me funciona

+++ Tema-2/Ejercicio-1/src/AplanarDirectorios.java
@@ -11,25 +11,21 @@ public class AplanarDirectorios {
         this.directorio = directorio;
     }
 
-    public void aplanar() {
-        try(Stream<Path> files = Files.walk(this.directorio).sorted(Comparator.reverseOrder())){
+    public void aplanar(){
+        try (Stream<Path> files = Files.walk(this.directorio).sorted(Comparator.reverseOrder())) {
             files.forEach(file -> {
-                if (Files.isDirectory(file)) {
-                    try {
-                        Files.delete(file);
-                    } catch (Exception e) {
-                        System.out.println("Error al borrar el directorio " + file);
-                    }
-                }else {
-                    try {
-                        Files.move(file, this.directorio.resolve(file.getFileName()));
-                    } catch (Exception e) {
-                        System.out.println("Error al mover el archivo " + file);
-                    }
+                try {
+                    if (!Files.isSameFile(file, this.directorio))
+                        if (Files.isDirectory(file))
+                            Files.delete(file);
+                        else
+                            Files.move(file.toAbsolutePath(), this.directorio.resolve(file.getFileName()));
+                } catch (IOException e) {
+                    System.out.println("Error al mover el archivo " + file.getFileName() + " al directorio " + this.directorio + ".");
                 }
             });
-        } catch (IOException e) {
-            throw new RuntimeException("Error al leer el directorio " + this.directorio);
+        }catch (IOException e) {
+            throw new RuntimeException("Error al leer el directorio " + this.directorio + ".");
         }
     }
 }
